% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morant.R
\name{fivethirtyeight_pal}
\alias{fivethirtyeight_pal}
\title{FiveThirtyEight color palette}
\usage{
fivethirtyeight_pal()
}
\description{
The standard three-color FiveThirtyEight palette for line plots comprises
blue, red, and green.
}
\seealso{
Other colour fivethirtyeight: 
\code{\link{scale_colour_fivethirtyeight}()}
}
\concept{colour fivethirtyeight}
