% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morant.R
\name{theme_morant}
\alias{theme_morant}
\title{Tema inspirado en nuestros colores institucionales}
\usage{
theme_morant(base_size = 18, base_family = "Futura")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Tema inspirado en nuestros colores institucionales
}
\concept{temas morant}
